requirements.txt
# PDF processing
pdfplumber==0.10.4
PyMuPDF==1.24.0

# Vector database
qdrant-client==1.7.3

# Embeddings and LLM
openai==1.35.3
anthropic==0.31.1
tiktoken==0.5.2
langchain==0.1.7
sentence-transformers==2.5.1

# Data processing
numpy==1.26.4
pandas==2.2.0

# Environment management
python-dotenv==1.0.1

# UI (optional)
gradio==4.19.1
streamlit==1.31.1

# Utilities
tqdm==4.66.2
colorama==0.4.6 